Reorganizing finalizer code into separate file
